requirements.txt
aiosqlite==0.20.0
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.3.0
babel==2.16.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blivet==3.10.0
blivet-gui==2.5.0
boto3==1.35.37
botocore==1.35.37
Brlapi==0.8.5
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cryptography==41.0.7
cupshelpers==1.0
Cython==3.0.9
dasbus==1.7
databases==0.9.0
dbus-python==1.3.2
distro==1.9.0
dnf==4.21.1
dnspython==2.6.1
email_validator==2.2.0
evdev==1.7.1
Faker==28.4.1
fastapi==0.114.1
fastapi-cli==0.0.5
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
grapejuice==3.40.14
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
humanize==3.13.1
idna==3.8
iniconfig==2.0.0
iso639==0.1.4
Jinja2==3.1.4
jmespath==1.0.1
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.3
louis==3.28.0
lutris==0.5.17
lxml==5.1.0
Mako==1.2.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moddb==0.11.0
nftables==0.1
numexpr==2.8.5
numpy==2.1.1
olefile==0.47
packaging==24.1
pandas==2.2.2
Paste==3.7.1
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
pluggy==1.5.0
ply==3.11
productmd==1.38
progressbar2==3.53.2
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pyenchant==3.2.2
pygit2==1.14.0
Pygments==2.18.0
PyGObject==3.48.2
PyGObject-stubs==2.11.0
pykickstart==3.52
pymongo==4.8.0
pyOpenSSL==23.2.0
pyparted==3.13.0
pypresence==4.3.0
pyproj==3.7.0
PySocks==1.7.1
pytest==8.3.3
python-augeas==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-meh==0.51
python-multipart==0.0.9
python-pam==2.0.2
python-utils==3.7.0
pytz==2024.2
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
rich==13.8.0
rpm==4.19.1.1
rpmautospec==0.7.2
rpmautospec-core==0.1.5
s3transfer==0.10.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==74.1.2
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.7.2
soupsieve==2.5
SQLAlchemy==2.0.36
starlette==0.38.5
systemd-python==235
Tempita==0.5.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
wheel==0.44.0
xkbregistry==0.3
zstandard==0.22.0
